refactor(cli): port --speech to shared voice loop via StdoutObserver
speech_loop.rs becomes a directory module with a thin run adapter
and a StdoutObserver that preserves the existing CLI print formatting
verbatim. State machine implementation now lives in
primer-speech::voice_loop and is shared with the GUI.

DialogueResponder + audio_thread (run_audio_thread + ChannelStt
glue) stay in primer-cli — they own the cpal-driven audio capture
thread which is a CLI-specific concern. The GUI will build its own
audio-thread plumbing in a later PR.

Cargo.toml: add primer-speech/voice-loop to the speech feature
so the shared state machine is compiled in alongside silero/whisper/
piper/cpal as before.
